fix: address code style issues
