+++ src/Service/PasswordResetService.php
@@ -116,4 +116,3 @@ class PasswordResetService
         $stmt->execute([(new DateTimeImmutable())->format('Y-m-d H:i:s')]);
     }
 }
-
