+++ tests/Service/PasswordResetServiceTest.php
@@ -6,29 +6,7 @@ namespace Tests\Service;
 
 use App\Service\PasswordResetService;
 use App\Service\UserService;
-use Psr\Log\AbstractLogger;
 use Tests\TestCase;
-
-class ArrayLogger extends AbstractLogger
-{
-    /** @var list<array{level:string,message:string}> */
-    public array $records = [];
-
-    public function log($level, $message, array $context = []): void
-    {
-        $this->records[] = ['level' => (string)$level, 'message' => (string)$message];
-    }
-
-    public function has(string $level, string $fragment): bool
-    {
-        foreach ($this->records as $r) {
-            if ($r['level'] === $level && str_contains($r['message'], $fragment)) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
 
 class PasswordResetServiceTest extends TestCase
 {
