+++ tests/bootstrap.php
@@ -2,3 +2,4 @@
 
 require __DIR__ . '/../vendor/autoload.php';
 require __DIR__ . '/TestCase.php';
+require __DIR__ . '/Service/ArrayLogger.php';
